never export emscripten generated invoke_ functions
Adding new extensions adds new functions we have to add manually.
The way I understand this is that we never want to expose any invoke_*
function any way so we can generally remove them.

This also contains a fix to make timescaledb work with pqlite. There we
need to expose pqsignal. But this is excluded via the libpq/exports.list
In this patch we explicitly remove it from the exclusion list.